netfs: Rename unlock_page_fscache() and move wait_on_page_fscache()
Rename unlock_page_fscache() to unlock_page_private_2() and change the
references to PG_fscache to PG_private_2 also.  This makes it look more
generic and doesn't mix the terminology.

Fix the kdoc comment on the above as the wake up mechanism doesn't wake up
all the sleepers.  Note the example usage case for the function in
conjunction with the cache also.

Place unlock_page_fscache() as an alias in linux/netfs.h.

Move wait_on_page_fscache() to linux/netfs.h.

[v2: Implement suggestion by Linus to move the wait function into netfs.h]

Reported-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: David Howells <dhowells@redhat.com>
Link: https://lore.kernel.org/linux-fsdevel/1330473.1612974547@warthog.procyon.org.uk/
Link: https://lore.kernel.org/linux-fsdevel/CAHk-=wjgA-74ddehziVk=XAEMTKswPu1Yw4uaro1R3ibs27ztw@mail.gmail.com/